@@ -1,7 +1,6 @@
-import { Injectable, Inject } from '@angular/core';
+import { Injectable } from '@angular/core';
 //
 import { MetaInfo, GenericFieldInfo, MetaInfoTag, ControlType } from '../meta-info/meta-info.model';
-import { MetaInfoExtraDataService } from './meta-info-extra-data.service';
 import { CrudConfig } from '../models/crud-config';
 
 @Injectable({
@@ -10,8 +9,7 @@ import { CrudConfig } from '../models/crud-config';
 export class MetaInfoBaseService {
   private metaInfoDefinitions: Map<MetaInfoTag | MetaInfoTag, MetaInfo>;
 
-  constructor(private metaInfoExtraDataService: MetaInfoExtraDataService,
-    private crudConfig: CrudConfig) {
+  constructor(private crudConfig: CrudConfig) {
     this.metaInfoDefinitions = this.crudConfig.metaInfoDefinitions;
   }
 
@@ -29,29 +27,21 @@ export class MetaInfoBaseService {
     return restPath;
   }
 
-  public extractRestPath(metaInfo: MetaInfo, data: any = null): string {
+  public extractRestPath(metaInfo: MetaInfo, parentKeyValue: any): string {
     if (metaInfo.restPath) {
       if (metaInfo.restPath.includes('{')) {
+
         const restPathArr = metaInfo.restPath.split('/');
         restPathArr.forEach((part: string, index: number) => {
           if (part.startsWith('{') && part.endsWith('}')) {
-            const extraDataKey = part.substr(1, part.length - 2);
-            const restPathId = extraDataKey.length ? this.metaInfoExtraDataService.getExtraData(extraDataKey) : null;
-
-            if (restPathId) {
-              restPathArr[index] = restPathId;
-            } else if (data) {
-              restPathArr[index] = data[extraDataKey];
-            } else {
-              restPathArr[index] = null;
-            }
+            restPathArr[index] = parentKeyValue;
           }
         });
         if (restPathArr.some((path: string) => path === undefined)) {
           return null;
+        } else {
+          return restPathArr.filter(Boolean).join('/');
         }
-
-        return restPathArr.filter(Boolean).join('/');
       } else {
         return metaInfo.restPath;
       }
@@ -72,6 +62,17 @@ export class MetaInfoBaseService {
       console.error(`Crud: No primaryKeyName found`);
     }
     return primaryKeyName;
+  }
+
+  public getPrimaryKeyValue(metaInfoSelector: MetaInfoTag, data: any): any {
+    const metaInfo = this.getMetaInfoInstance(metaInfoSelector);
+    const primaryKeyName = metaInfo?.fields?.find(field => field.isPrimaryKey)?.name;
+    if (!primaryKeyName) {
+      console.error(`Crud: No primaryKeyName found`);
+      return null;
+    } else {
+      return data?.[primaryKeyName];
+    }
   }
 
   public getMetaInfoInstance(metaInfoSelector: MetaInfoTag): MetaInfo {
